feat: Add animated sprites for heroes
Add animated sprites for heroes 7, 8, 9, 10, 11, 12, 13, 14, and 15. The sprites have been scaled to 4x4 and the idle animation has been set as the default animation.

feat: Update battleground and hero manager

In the battleground script, add the _ready() and _process() functions. In the hero manager script, add a unit test for the hero manager.

feat: Update animated sprite properties

Update the scale property of the animated sprites for heroes 7, 8, 9, 10, 11, 12, 13, 14, and 15 to be 4x4.

feat: Update farm scene

Update the farm scene to use the new Serene Village tileset.

feat: Add campfire and door sprites

Add the campfire and door sprites to the Serene Village tileset.

refactor: Remove unused files

Remove the Welcome.md file and the inn.png.import file.

test: Add unit tests for Gut plugin

Add unit tests for the Gut plugin and update the Gut configuration files.

test: Add unit test for MoneyManager

Add a unit test for the MoneyManager script.

test: Add unit test for HeroManager

Add a unit test for the HeroManager script.